android: loads firmware and fpga onto B200 if they are not already there.
The program has to be run twice to actually load both because the
firmware changes how the system recognizes the device, so it has to
re-ask permission to use it over USB. So the first run loads the
firmware and the second run will ask for USB permission again and then
load the FPGA image. Once both are loaded, the app will start as usual.

Need to provide some feedback to the application for the percentage
when loading firmware and fpga.